% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/felp.R
\name{felp}
\alias{felp}
\title{Returns source and help of a function simultaneously}
\usage{
felp(topic, package = NULL, ...)
}
\arguments{
\item{topic}{A name or character string specifying the function for which help and source are sought. Package to be sought can also be defined simultaneously if x is specified in the style of package::name or package:::name}

\item{package}{A name or character string specifying the package for which help and source are sought. If the package is specified by x, this parameter is neglected, or NULL (default).}

\item{...}{Other arguments passed to help}
}
\description{
Returns source and help of a function simultaneously
}
