% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.function}
\alias{print.function}
\title{Display source and help of function}
\usage{
\method{print}{function}(x, ...)
}
\arguments{
\item{x}{function}

\item{...}{Other arguments passed to felp. felp is called when print.function is explicitly called or via print.}
}
\description{
Display source and help of function
}
