% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telp.R
\name{telp}
\alias{telp}
\title{Translated Help}
\usage{
telp(topic, package = NULL, translate_to = "default", ...)
}
\arguments{
\item{topic}{usually, a \link{name} or character string specifying the
   topic for which help is sought.  A character string (enclosed in
   explicit single or double quotes) is always taken as naming a topic.

   If the value of \code{topic} is a length-one
   character vector the topic is taken to be the value of the only
   element.  Otherwise \code{topic} must be a name or a \link{reserved}
   word (if syntactically valid) or character string.

   See \sQuote{Details} for what happens if this is omitted.
 }

\item{package}{a name or character vector giving the packages to look
   into for documentation, or \code{NULL}.  By default, all packages
   whose namespaces are loaded are used.  To avoid a name being deparsed use e.g.
   \code{(pkg_ref)} (see the examples).}

\item{translate_to}{The help document will be translated to the specified
language (e.g., \code{"ja"} for Japanese and \code{"de"} for German). If \code{"default"},
the function looks up \code{getOption("telp_translate_to", "en")}.}

\item{...}{Arguments passed to \code{utils::help}}
}
\description{
The translated help is available when \code{topic} belongs to a \code{package} installed
from CRAN and the system is online. Otherwise, normal help will be shown.
The source of the help is powered by 'RDocumentation', and
the translation is powered by 'Google Translate.
}
\examples{
\dontrun{
telp(help)
}

}
