% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzyhelp.R
\name{fuzzyhelp}
\alias{fuzzyhelp}
\title{Fuzzily Search Help and View the Selection}
\usage{
fuzzyhelp()
}
\description{
Users no more have to afraid of exact name of the object they need help.
A shiny gadget helps you to find a topic fuzzily.
Click radio buttons to switch preview contents.
Click "Done" or "Cancel" to close the widget.
The "Done" button will also hook \code{help} function on the selection.
}
\note{
The fuzzy match algorithm is experimental, and may change in the future.
}
