% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{?}
\alias{?}
\alias{?.function}
\alias{?.default}
\title{S3 version of \link[utils:`?`]{utils::?}}
\usage{
"?"(e1, e2)

\method{?}{function}(e1, e2)

\method{?}{default}(e1, e2)
}
\arguments{
\item{e1}{A topic ought to be documented. Refer to \code{topic} argument described in \code{utils::`?`}.}

\item{e2}{\code{.} equals to \code{missing(e2)}. If else, see \code{type} argument described in \code{utils::`?`}.}
}
\description{
S3 version of \link[utils:`?`]{utils::?}
}
