% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{?}
\alias{?}
\alias{?.function}
\alias{?.default}
\title{S3 version of [utils::`?`]}
\usage{
"?"(e1, e2)

\method{?}{function}(e1, e2)

\method{?}{default}(e1, e2)
}
\arguments{
\item{e1}{A topic ought to be documented. Refer to `topic` argument described in `` utils::`?` ``.}

\item{e2}{`.` equals to `missing(e2)`. If else, see `type` argument described in `` utils::`?` ``.}
}
\description{
S3 version of [utils::`?`]
}
