% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{?}
\alias{?}
\title{Functional help with \code{?} operator}
\usage{
"?"(e1, e2)
}
\arguments{
\item{e1}{A topic of document.
Refer to \code{topic} argument described in \code{utils::`?`}.}

\item{e2}{\code{.} and \code{p} have special meanings as documented above.
Otherwise, \code{e2} is same as \code{type} argument of \code{utils::`?`}.}
}
\description{
Displays help and structure of an object, or help of a package.
Two syntax sugars are added to enhance \code{utils::?}.
One is \code{object?.} which works as if \code{?object}.
Another is \code{package?p} which works as if \code{help(package = package)}
}
\examples{
# Identical to help(identity); print(identity)
?identity
identity?.  # The same

# Identical to help(iris); str(iris)
?iris
iris?.  # The same

# Identical to help(package = MASS)
MASS?p

}
